First draft of the am and init compose file and
This is the first take on the am compose file.

Also I added an init compose file with which it is possible to create
all the certificates for the AM and also the initial configuration for
the am database and PAP.

For the init container to work we need the files in config/
These are templates which will be used by the init coantiner to create
the real configuration/setup for the am container.